styled and validated section
