+++ src/components/email-sign-up/styles.js
@@ -1,0 +1,10 @@
+const styles = {
+  emailSignUpContainer: {
+    marginTop: 20
+  },
+  signUpButton: {
+    marginLeft: 10
+  }
+};
+
+export default styles;
