+++ src/containers/add-interest/index.js
@@ -2,6 +2,7 @@ import React, { Component } from "react";
 import PropTypes from "prop-types";
 import { Typography } from "@material-ui/core";
 import CardContainer from "../../components/card-container";
+import SubTitle from "../../components/sub-title";
 import EmailSignUp from "../../components/email-sign-up";
 import { withDb, addInterestedUser } from "../../libs/db";
 
@@ -21,10 +22,12 @@ class AddInterest extends Component {
   render() {
     return (
       <CardContainer>
-        <Typography variant="h2">Want to know when we are ready?</Typography>
-        <Typography variant="p">
-          Sign up with your email and you will be in line to try us out!
+        <Typography variant="h2" gutterBottom>
+          Want to know when we are ready?
         </Typography>
+        <SubTitle>
+          Sign up with your email and you will be in line to try us out!
+        </SubTitle>
 
         <EmailSignUp onButtonClick={this.submitEmail} />
       </CardContainer>
